Fix OCMB scom path in HBRT testcases
At runtime in HBRT, the scom path for the OCMBs (both Explorer
and Odyssey) is diffferent from the processor or previous memory
buffers.  Instead of a scom address being passed in, the complete
mmio address is used as the argument.  The previous implementation
of the HBRT backend for our standalone simics testcases was
calling into our internal SCOM driver.  This caused some confusion
with the input address as well as leading to some duplicate code
paths being executed.  The change here is to call into our MMIO
driver instead which more closely matches how PHYP would handle
the call.

Change-Id: I806b7978fd7acfaf987d67f85bdb929fbf07509a
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/174787
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>