New guard type for spare cores
There is a requirement from PE to differentiate the guard records
that are applied against spare cores.  To achieve this we will
create a new guard type.

Hostboot will use this new type in two places:
1- Any time we lower the severity of the log due to remaining spares
2- When applying guard records during the IPL if there are spares
   remaining.  This will handle the checkstop case where BMC/FSP
   creates the record but have no sense of spares.

The details of 1 will be covered when the severity-lowering logic
is added. 2 is covered here along with the definition of the new
type.

As part of this support, a generic function was created that can
be used for all of the situations where a core/fc error is detected
and we want to try to swap in the spare.  This function will check
and update the live attributes that note spare deployment.

Change-Id: If28dd05c3d0b42e320f38a5f471785917bf314e3
JIRA: PFHB-810
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/176611
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>