std library fixes: Use move instead of copy
Updates std::remove_if and std::vector<>.erase to use move instead of
copy which brings the code more in line with C++ std library. This
allows a vector of unique_ptrs to be modified where otherwise a compile
failure would occur because the code would try to copy-assign a unique_ptr
which is not allowed.

Ex.

std::vector<unique_ptr<int>> a;
// fill a with stuff
a.erase(std::remove_if(begin(a), end(a),
    [](int i){ return i == 3;}), end(a));

Change-Id: Ib84302c2a82f5bb560dd1a8fed3b7fc7b3bc4801
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/176934
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: MIKE VAGELINE <mpvageli@us.ibm.com>