Track spare core deployment in attributes
Creating 2 new attributes to track when spare cores are deployed.

ATTR_REPLACED_BY_SPARE will be set on a core when it is found to
be at fault such that an available spare will be used to take its
place.

ATTR_SPARE_CORES_DEPLOYED will be incremented any time a core
fault is detected and a spare is deployed.  Note that this number
should always match the total number of cores on this chip that
have ATTR_REPLACED_BY_SPARE set.  Once this attribute matches
ATTR_SPARE_CORES, there are no more spares available to deploy.

Change-Id: Ifcd5b83beb0054f328108e3d4181e40e93826bf5
JIRA: PFHB-810
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/177113
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>