Update DCE makefile to filter out main file
The main DCE file was erroneously being passed to the compile when a
user kept their main file in the same location as their extra files.
This resulted in multiple definition errors.

This commit fixes that issue by filtering out that file if it is present
in the extra files location. This allows users to store their main file
in the same location as their extra files.

A practical use of this would be having the extra file location pointed
at a place in the src tree where the main file also lives. For example

src/usr/hdat/main-dce.C

By having the main in the same location, you can pick up all the include
paths for that directory for free (i.e. not having to change dce_rc
incdirs) when you compile and you can modify any of the existing hostboot
files in that location as "extra files" for you dce program. All without
having to 'cd' to different places or remember file paths.

Change-Id: I7942163d4358921f6a6cde950647ae43fbdec536
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/177708
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>