Hostboot workaround to get visibility to ody pll unlock errors
This commit is used to address the STG Defect 640599.

The requirements from the defect are as follows:
PLL unlock FIR (TPC FIR bit 18) is currently masked in FW.
We want the following so that it can go into 1060 service pack.

1. Need hostboot to unmask (bit 12 of 0x010F001E),
   since not allowed to make 1060 SBE updates.
2. Want to unmask as early as possible in the IPL
   (as early as ody_sppe_check_for_ready ?).
3. Want to unmask as part of concurrent code update.

This commit tries to unmask bit 12 of chiplet 0x010F001E in
istep 12.1.

STGD: 640599
Change-Id: Ibc3e193dd7cf60033e267be4b0495a334b2c0ff4
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/176995
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>