Reduce severity of logs with spareable unit callouts
Adds code in the commit path for error logs to check if the error being
committed has spareable units in its callouts. When there are still
spares available to deploy the error will have its severity reduced and
the spare units will have any guards changed to GARD_Spare

An exception to this is if there are multiple service actions or if
there are callouts with higher priority than the core/fc callout.

Change-Id: Iebd15d72c5786ce9db31d237dbcab3c766d29df7
JIRA: PFHB-628
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/179559
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>