Update secureboot securerom interfaces for V3 support
Since secureboot V1 and V3 security modes need to be supported in
the same build there will be times where we need to override the
default of using whatever mode the system is using.  This commit
updates the interfaces to allow the use of V1 or V3 modes
where appropriate.

JIRA: PFHB-676
Change-Id: I15eba2fae7e92571b0fea81a16f940bf48bc60ce
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/179854
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>