Allow service actions for non-visible logs when SRC bits are set
For spareable errors, we want service actions to still take place even
if the log is made non-visible so that the core/fc can be knocked out.
Before, SRC bits were not switched off if the log was non-visible but
can now be used to allow service actions on non-informational logs.

Change-Id: I1d60f509020560b833fae53536333eed7db79699
JIRA: PFHB-629
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/179560
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>