Hostboot Secureboot V3 Build Support - Pass 2
This commit updates genPnorImages.pl and some supporting files
to securely sign all of the files in a pnor layout xml file
with both V1 (existing) and V3 (new) security algorithms.

The new V3 files are located in the sub-folder "V3" of
the pnor or staging directory used for the V1 files.

JIRA: PFHB-785
Change-Id: I3e02f0082ff8b49b6a1e076380c6bd15f6949774
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/179857
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>