Fix early fail module dependency in trace
The tracedaemon module is loaded after error logs could be
created.  The collectThreadTrace call had a dependency on
a function that was part of libtracedaemon.  This caused
an Instruction Storage Exception when logged an early
failure from the pnor logic.  The fix is to move this
object into the libtrace module that is resident at
this point.

Change-Id: I15194798172cea884837e9ebab168c8746192cb7
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/180444
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>