Create HB_HLL PNOR Section/LID
This commits creates the HB_HLL PNOR Section, aka LID.  The
HB_HLL stands for Hostboot Hash Lid List, and it is based off
of PHYP's new HLL lid.

The HB_HLL contains the V3 hashes for the other PNOR sections
(lids) that hostboot will have to verify.  It contains an HLL
Header (aka TOC) that designates how many V3 hash entries are
stored in the HB_HLL.

Also, the HB_HLL is signed with the V3 secureboot algorithm.
The V3 15KB secureboot header created from the signing is
appended to the start of the partition, before the HLL Header
section mentioned above.

While this commit covers the creation of the HB_HLL, how it is
consumed by Hostboot code will be covered in another commit.

JIRA: PFHB-786
Change-Id: Ia1c6d918a50b3f5bab4b13110868bdcd30ecc3b2
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/180608
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>