Saving and resetting buffer length after retries in Scom.C
Changes:

-- scom.C:
In doScomOp() I added old_io_buflen as a way to perserve the passed
in buffer length in the case that it is set to 0 or another value
during retry looping and an error is not raised.

This was done to hopeully address the remaining issues related
to the false paritity error that occasionally would appear during
runtime.

Change-Id: Icc20bda18c0e08ed00986e16d31bdff8713c60f3
STGD:662123
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/181440
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Connor Leu <connor.leu@ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>