Update the Secureboot Signing Mode during Processor SBE Update
This commit updates the Secureboot Signing Mode that is used
when customizing the SBE image for a processor.  This commit
defaults the signging mode to the system's signing mode, but
adds ATTR_SB_SIGNING_MODE_OVERRIDE so that the value can be
overridden for testing purposes.  It also adds a check such
that it will not use ATTR_SB_SIGNING_MODE_OVERRIDE if
booting on a production driver.

JIRA: PFHB-685
Change-Id: I9cc485ae9cbf43d0d240b2c7dbbab559ac342a90
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/181513
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Chris Engel <cjengel@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>