Fix bug in call_mss_getecid.C
The call_mss_getecid() loops through a list of Odyssey OCMB targets
and does scom read and writes to the CHIPLET1_REGISTER to unmask
bit 12. However, the logic makes use of a numBytes variable which
is initialized to a value of 8. However, if any of the scom
operations fail, there is a potential for this value to be set
to 0 and this could cause the subsequent scom operation on the
next OCMB to fail due to invalid num of bytes passed for
the scom operation!

We saw this issue with defect 665806 where one of the OCMBs
had downlevel firmware and the scom operation had failed and that
went down the code update path, but recovered. However the earlier
failed operation caused the next OCMB's scom operation to fail
as the failed scom read had set the length to 0.

The fix is to explicitly set the numBytes to 8 before each
scom read and write operation to ensure we are passing the
correct value.

STGD: 665806
Change-Id: I29f7bc9e1805cb67fd2cf31c041741225ecf016f
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/182253
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>