Add HLL to the PreVerified Lid List
When the system is using V3 secureboot signatures, the HLL (Hash
List Lid) is used to know what lids need to be loaded for
PHYP (PowerVM) itself and also lids that need to be PreVerified
for PHYP to use when it starts.  This commit adds the HLL itself
to the PreVerified list so that PHYP can start with immediate
access to it.

This commit also updates some security-related traces and console
messages that hostboot make earliy in the IPL to better capture the
system's security settings.

JIRA: PFHB-680
Change-Id: Iad6e8465ccfa1ab8625f05f81dce792878813f06
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/182471
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Victor Umoren-Udo <umorenudovictor@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>