PRD: fix assert on FSP checkstop that fails to isolate
An assert failure will happen if there is a system checkstop on an FSP
machine and PRD isolates to a unit of a chip, but that unit has no
active attentions. This will result in an FSP reset, but because the
checkstop attention has not been cleared PRD will find the same scenario
after the reset thus causing an infinite loop (at least until the
watchdog timer expires).

The root of the problem is that the "no active attentions" triggers a
default callout. That sends us down some error path handling in
RasServices that will try to get the node target from the unit target,
which is not possible (at least not directly). The solution is to hop
from the unit to the parent chip target. Then get the node target from
the chip target.

Change-Id: I67f026f570ad272b97cdd104f127a39791e0f863
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/prd/+/182419
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Zane C Shelley <zshelle@us.ibm.com>
Reviewed-by: Caleb N Palmer <cnpalmer@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/184601