Fix spare guard creation for HBRT
There was an issue where when a spare error was detected at runtime the
error manager would take a different path to process callouts in an
error log such that the code would back out before the guard record was
created. This caused PHYP to never get notified that the spare should
be deployed.

This commit addresses a few places in the code where it would exit early
for non-visible logs.

Change-Id: I13a805bbb8fa1ae1350c2078db4fe0bfe2bff880
STGD:667682
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/182662
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Veena Ganti <veena@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>