Skip PLL Unlock FIR modification on downlevel parts
The level of Odyssey SBE code that is currently being
installed by the DDIMM vendors is of a pre-GA vintage.
Specifically this level of code is missing a change to
the allow-deny list related to FIRs.  A change added
post-GA does a scom that is blocked by this level of
code.  This error is recoverable by doing a sbe code
update, however since it happens in step 12.1 we are
forced to do the update via i2c.  It takes 2.5 minutes
to update a single SBE which adds up to several hours
for a fully loaded system.

The workaround for this issue is to explicitly check
for this known level of downlevel SBE code and skip
the explicit scom.  This will only be done when this
is a manufacturing ipl.

Change-Id: I9dda7fbff3e12343bbfb21aa44be2c32abab17db
STGD: 668430
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/182742
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>