Fix double commit in Odyssey error path
The ody_upd_process_event() call takes in an error log that
always gets committed.  However, since it is not passed by
reference the caller does not know that happens.  This leads
to the caller also committing the log again later.  This causes
memory corruption due to a double-delete of the contents
and/or accessing of memory that has already been deleted.

Change-Id: Ie2c602d72f5f3a274fcd55e1145bdda831c808dd
STGD: 671228
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/183766
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>