Append V3 Header as "unprotected" data to HBBL
For FW1110 Hostboot code needs to support both V1 and V3
security headers in the code base.  Specifically, for the
Hostboot Bootloader (HBBL) both the V1 security header and
the V3 security header are needed to customize the
Processor SBE Image in a future story.

This commit generates the V1 and V3 headers for the HBBL
binary after it has been padded to a page (4KB) boundary:
[V1 Header][HBBL data + padding to page boundary][V3 Header]

This will make the HBBL look almost like a PNOR partition with
a hash page table.  In this case the HBBL binary + padding is
the "protected" data (like the hash page table) and the V3 header
is the "unprotected" data.

This commit also updates the PNOR layout files to accomodate the
larger HBBL size.

JIRA: PFHB-786
Change-Id: Idc75f0370ca1d198ef07915904339e1f7529403e
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/184177
Reviewed-by: MIKE VAGELINE <mpvageli@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Chris Engel <cjengel@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>