Add HBBL V3 Secureboot Header to Processor SBE Update
During a processor SBE Update both a truncated version of the
Hostboot Bootloader's (HBBL) V1 security header and the full
version of V3 security header need to be added to the .sbh_hbbl
XIP section of the customized SBE Image.

The truncated V1 header is already being added. This commit handles
the interfaces required to add the V3 header.  This includes the
PNOR interfaces required to point to the V3 header in memory, as it
is now the "unprotected" part of the HBBL pnor section.

This commit also improves the memory usage during the update
procedure by saving off the HBBL data and the HBBL V1 and V3
security headers such that each thread can access them without
having to create the memory space itself.  Some testcases are also
updated to support this improved memory usage, as well.

JIRA: PFHB-685
Change-Id: Ic6cbeca541374daf77770cac9283df152f8b0283
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/184351
Reviewed-by: Chris Engel <cjengel@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>