Improve HBBL trustedboot actions in MPIPL
In MPIPL when there are multiple TPMs only one will be non-poisoned and that is
the primary TPM. The bootloader does not know which TPM is the primary without
some investigation which it cannot do presently. To simplify the hash extend
logic so that the primary TPM always gets the hash, the bootloader will find
all TPMs and do a hash extend on them regardless if they were poisoned or not.

For BMC systems, there is only one TPM so this logic will only extend to the
sole TPM.

For FSP systems, there are more than one TPM and more than one boot chip.
The bootloader will extend to all TPMs.

Change-Id: I6c649278f783637351d426bc7f689e6bc1bd07bc
JIRA:PFHB-784
STGD:620212
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/183839
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>