TMGT: Send full mem power when preheat power not available
Currently the full mem power was only sent when HTMGT was sending the
preheat memory power data. This commit will always send the full mem
power so the OCC can calculate the memory power.

Also change the default utilization points to 30% and 50%.
(This change will only be picked up on new systems on initial run of
serverwiz)

Change-Id: Id6256cb5fa72381cdd21357b84c8046f684c0e0d
JIRA: PFES-95
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/184970
Reviewed-by: Martha Broyles <mbroyles@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Sheldon Bailey <baileysh@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>