Add new attribute to control SCOM traces
There is currently no way to enable SCAN/SCOM traces on a BMC
system without recompiling Hostboot code because attribute
overrides don't work on complex attributes.  This change adds
a new independent attribute alongside the existing HB_SETTINGS.

There is also an order-of-operations issue that had to be
resolved.  The attribute overrides are not applied until
after the trace code checks the values.  This change then also
adds an external interface that can be called to reevaluate
the trace settings to handle any that were overridden.

Change-Id: Icad9a2bfafb75c63434a5147c55b51bd97e7f26d
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/185768
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: MIKE VAGELINE <mpvageli@us.ibm.com>
Reviewed-by: Daniel Crowell <dcrowell@us.ibm.com>