Add V3 Algorithms to ContainerHeader class in Hostboot Code
This commit adds support for V3 container headers to the existing
support of V1 container headers in hostboot's ContainerHeader class.

This allows Hostboot code throughout the IPL - from early PNOR and
Secure PNOR calls to later "prepare for runtime" calls - to handle
the existence of both V1 and V3 container headers in the same build.

Additionally, this commit securely loads the HB_HLL partition in a
testcase which is the only partition that currently starts with a
V3 container header in each build.  This verifies the "good path"
of the V3 support alongside the existing V1 support.  However,
this "load" will only happen when/if the system signing mode changes
from V1 to V3 signing mode in a latter commit.

Change-Id: I8ade22e720c4c5d85a4d55e5020baa26c8fc129b
JIRA: PFHB-676
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/185859
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matthew Strom <matt.strom@ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>