V3 Cleanup: Change software references to firmware
In the past, there was a rename done in the spec for V1 security to
change references to "software" to "firmware" this was never reflected
in the code and can lead to confusion. This commit renames those old
references to bring the code in line with the specs.

Change-Id: I9857256fb2cdb02b85d5ae5eecb165893838219d
JIRA: PFHB-802
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/186636
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: MIKE VAGELINE <mpvageli@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>