Append V3 Header to the Hostboot Base Image (HBB)
For FW1110 Hostboot code needs to support both V1 and V3
security headers in the code base.  Specifically, for the
Hostboot Base Image (HBB) both the V1 security header and
the V3 security header are needed so that the Hostboot
Bootloader (HBBL) can verify the HBB image loaded by the
service processor depending on what signing mode (V1 or
V3) the system is in.

This commit generates the V1 and V3 headers for the HBB
binary after it has been padded to a page (4KB) boundary:
[V1 Header][HBB data + padding to page boundary][V3 Header]

This will make the HBB look almost like a PNOR partition with
a hash page table.  In this case the HBB binary + padding is
the "protected" data (like the hash page table) and the V3 header
is the "unprotected" data.

JIRA: PFHB-679
Change-Id: I27ee2a3e4038b2f3fd72b8df02b53d02637fc922
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/187160
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matthew Strom <matt.strom@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>