Add flexibility to call_rom_v3_verify() function
This commit adds an optional additional parameter to the
call_rom_v3_verify() routine that specifies the start of the
data that the container refers to.

If nullptr (which is the default value) is passed in for the
data offset the function will look for the data at the expected
location of the memory just past the V3 header at the start of
the container.  Otherwise, the code will use the memory address
passed in by the caller.

JIRA: PFHB-677
Change-Id: I94a1073054eee1d212704ef3e88618364b779932
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/187168
Reviewed-by: Matthew Strom <matt.strom@ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>