p10_adu_utils -- assert OVERWRITE_PBINIT for hotplug operations
Current code sets the OVERWRITE_PBINIT bit only when attempting
to issue an explicit pbop.enable_all command (as generally the fabric
is quiesced and this operation is required to leave that state, we
want to ensure the operation can be issued when pb_init is low)

This defect shows the ADU effectively getting stuck in the midst
of executing a hotplug sequence, where a pmisc (switch) commmand
is "wrapped" between a pbop.disable_all and pbop.enable_all.  In
the dry case, the pmisc command will issue without OVERWRITE_PBINIT
being set before triggering the hotplug sequence.  In the case the
pmisc command is retried, the ADU appears to get stuck however
(resampling pb_init and finding it low, it will wait forever).

This change sets OVERWRITE_PBINIT for all hotplug cases to avoid
this undesriable behavior.

Change-Id: Ia90f3c385833f009c5a13cf250589214304ec440
STGD:688176
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/ekb/+/187995
Reviewed-by: Thi N Tran <thi@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: HWSV CI <hwsv-ci+hostboot@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/188001
Tested-by: Daniel Crowell <dcrowell@us.ibm.com>
Reviewed-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel Crowell <dcrowell@us.ibm.com>